Force stubs from chibios-syscalls
by dropping the weak attribute. Also required
fixing the malloc locks to enfore initialization.
And revealed issues with the build system of
all the boards, since we were building syscalls
from ChibiOS as well as our own implementations
from chibios-syscalls or in the board firmware
directly. All fixed here.
