+++ can-io-firmware/Makefile
@@ -138,9 +138,9 @@ INCDIR = $(STARTUPINC) $(KERNINC) $(PORTINC) $(OSALINC) \
 
 include app_src.mk
 
-include $(CHIBIOS)/os/various/cpp_wrappers/chcpp.mk
+# ChibiOS C++ bindings
 CPPSRC += $(CHIBIOS)/os/various/cpp_wrappers/ch.cpp
-INCDIR += $(CHCPPINC)
+INCDIR += $(CHIBIOS)/os/various/cpp_wrappers
 
 UAVCAN = ../lib/uavcan
