+++ lib/chibios-syscalls/cpp_stubs.cpp
@@ -11,18 +11,18 @@ int __cxa_guard_acquire(__guard*);
 void __cxa_guard_release(__guard*);
 void __cxa_guard_abort(__guard*);
 void __cxa_pure_virtual();
-void* __attribute__((weak)) __dso_handle = NULL;
+void* __dso_handle = NULL;
 void _exit(int status);
 pid_t _getpid(void);
 int _kill(int pid, int sig);
 }
 
-void __attribute__((weak)) __cxa_pure_virtual()
+void __cxa_pure_virtual()
 {
     osalSysHalt("Pure virtual function call.");
 }
 
-void __attribute__((weak)) _exit(int status)
+void _exit(int status)
 {
     (void)status;
     osalSysHalt("Unrealized");
@@ -30,14 +30,14 @@ void __attribute__((weak)) _exit(int status)
     }
 }
 
-pid_t __attribute__((weak)) _getpid(void)
+pid_t _getpid(void)
 {
     return 1;
 }
 
 #undef errno
 extern int errno;
-int __attribute__((weak)) _kill(int pid, int sig)
+int _kill(int pid, int sig)
 {
     (void)pid;
     (void)sig;
