+++ master-firmware/package.yml
@@ -16,6 +16,7 @@ depends:
     - goap
     - quadramp
     - aversive
+    - chibios-syscalls
 
 target.arm:
     - src/main.c
@@ -42,7 +43,6 @@ target.arm:
     - src/can/can_io_driver.cpp
     - src/can/sensor_handler.cpp
     - src/can/uwb_position_handler.cpp
-    - src/malloc_lock.c
     - src/usbconf.c
     - src/aversive_port/rs_port.c
     - src/aversive_port/cs_port.c
