+++ motor-control-firmware/src/chconf.h
@@ -207,7 +207,7 @@
  * @note    The default is @p FALSE.
  * @note    Requires @p CH_CFG_USE_MUTEXES.
  */
-#define CH_CFG_USE_MUTEXES_RECURSIVE FALSE
+#define CH_CFG_USE_MUTEXES_RECURSIVE TRUE
 
 /**
  * @brief   Conditional Variables APIs.
