+++ proximity-beacon-firmware/Makefile
@@ -12,7 +12,7 @@ endif
 
 # C specific options here (added to USE_OPT).
 ifeq ($(USE_COPT),)
-  USE_COPT = 
+  USE_COPT =
 endif
 
 # C++ specific options here (added to USE_OPT).
@@ -29,7 +29,7 @@ endif
 
 # Linker extra options here.
 ifeq ($(USE_LDOPT),)
-  USE_LDOPT = 
+  USE_LDOPT =
 endif
 
 # Enable this if you want link time optimizations (LTO)
@@ -95,7 +95,6 @@ include $(CHIBIOS)/os/rt/rt.mk
 include $(CHIBIOS)/os/common/ports/ARMCMx/compilers/GCC/mk/port_v7m.mk
 include $(CHIBIOS)/os/hal/ports/STM32/STM32F3xx/platform.mk
 # include $(CHIBIOS)/test/rt/test.mk
-include $(CHIBIOS)/os/various/cpp_wrappers/chcpp.mk
 
 ifeq ($(USE_BOOTLOADER),)
   USE_BOOTLOADER = yes
@@ -111,7 +110,7 @@ else
   LDSCRIPT = $(BASE_DIR)/linker/STM32F303xC.ld
 endif
 
-include src/src.mk
+include src/app_src.mk
 
 # C sources that can be compiled in ARM or THUMB mode depending on the global
 # setting.
@@ -123,8 +122,7 @@ CSRC = $(PORTSRC) \
        $(STARTUPSRC) \
        $(CHIBIOS)/os/hal/lib/streams/chprintf.c \
        $(CHIBIOS)/os/hal/lib/streams/memstreams.c \
-       $(CHIBIOS)/os/various/syscalls.c \
-	   src/version.c \
+       src/version.c \
        $(PROJCSRC)
 
 # C++ sources that can be compiled in ARM or THUMB mode depending on the global
@@ -159,6 +157,10 @@ INCDIR = $(PORTINC) $(KERNINC) $(TESTINC) \
          $(HALINC) $(OSALINC) $(PLATFORMINC) $(STARTUPINC) \
          $(CHIBIOS)/os/various $(CHIBIOS)/os/hal/lib/streams $(CHCPPINC) \
          $(BASE_DIR)/board $(CHIBIOS)/os/license
+
+# ChibiOS C++ bindings
+CPPSRC += $(CHIBIOS)/os/various/cpp_wrappers/ch.cpp
+INCDIR += $(CHIBIOS)/os/various/cpp_wrappers
 
 #
 # Project, sources and paths
