+++ uwb-beacon-firmware/src/main.c
@@ -2,6 +2,7 @@
 #include <hal.h>
 #include <chprintf.h>
 #include <parameter_flash_storage/parameter_flash_storage.h>
+#include <chibios-syscalls/malloc_lock.h>
 
 #include "main.h"
 #include "usbconf.h"
@@ -29,6 +30,7 @@ void __late_init(void)
     /* C++ Static initializer requires working chibios. */
     halInit();
     chSysInit();
+    malloc_lock_init();
 }
 
 int main(void)
